g_app Server/requirements.txt
bootstrap-flask==2.2.0
datetime==5.5
flask==2.3.2
flask-cors==4.0.1
flask-socketio==5.3.6
eventlet==0.36.1
pyjwt==2.8.0